backend/requirements.txt
Django==3.2.9
graphene-django==2.15.0
django-graphql-jwt==0.3.0
django-graphql-auth==0.3.16
django-cors-headers==3.10.1
